Small corrections before PR.
